fix(models): 统一 21 个模型主键类型为 BigInteger (T-A-04)
将所有模型 id 主键字段从 Integer 改为 BigInteger，与外键类型保持一致，消除 openGauss 严格类型检查下的外键约束失败风险。SQLite 通过 BigInteger().with_variant(Integer, sqlite) 兼容。

对应任务：T-A-04（openGauss 适配阶段）

+++ backend/app/models/post.py
@@ -8,7 +8,7 @@ from app.database import Base
 class Post(Base):
     __tablename__ = "posts"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     school_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("schools.id"), nullable=False, index=True)
     category_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("categories.id"), nullable=False, index=True)
